+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 All notable changes to this project will be documented in this file. This
 project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+### Added 
+- Added `cu::Device::getArch()`
+
 ## \[0.8.0\] - 2024-07-05
 
 ### Added